+++ CHANGELOG.md
@@ -144,6 +144,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* Adds a new annotate command to annotate files line by line. This is similar
+  in functionality to git's blame. Invoke the command with `jj annotate <file_path>`.
+  The output can be customized via the `templates.annotate_commit_summary`
+  configi variable.
+
 * Support background filesystem monitoring via watchman triggers enabled with
   the `core.watchman.register_snapshot_trigger = true` config.